Client/src/main/java/Message.java
public class Message {

    private String sequence;
    private String timestamp;
    private String fromId;
    private String toId;
    private String message;

    public Message() {}

    public Message(String sequence, String fromId, String toId, String message) {
        this.sequence = sequence;
        this.fromId = fromId;
        this.toId = toId;
        this.message = message;
    }

    public void setTimestamp(String timestamp) {
        this.timestamp = timestamp;
    }

    public String getTimestamp() {
        return timestamp;
    }

    public String getSequence() {
        return sequence;
    }

    public void setSequence(String sequence) {
        this.sequence = sequence;
    }

    public String getFromId() {
        return fromId;
    }

    public void setFromId(String fromId) {
        this.fromId = fromId;
    }

    public String getToId() {
        return toId;
    }

    public void setToId(String toId) {
        this.toId = toId;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

}
